feat: refactor failures
