+++ Sources/DesignTokensCore/Errors/AliasValueFailure.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+/// The possible failures related to alias values.
+enum AliasValueFailure: ValueFailure {
+  /// The value of the alias is not a valid token reference.
+  case invalidReferenceSyntax
+}
