+++ Sources/DesignTokensCore/Errors/DecodingFailure.swift
@@ -8,31 +8,14 @@ enum DecodingFailure: Error {
   /// The token is missing a type.
   case missingType
 
+  /// The decoded color value is invalid.
   case invalidColorValue(tokenName: String, tokenPath: [String], valueFailure: ColorValueFailure)
   
+  /// The decoded dimension value is invalid.
   case invalidDimensionValue(tokenName: String, tokenPath: [String], valueFailure: DimensionValueFailure)
   
+  /// The decoded alias value is invalid.
   case invalidAliasValue(tokenName: String, tokenPath: [String], valueFailure: AliasValueFailure)
 }
 
 extension DecodingFailure: Equatable {}
-
-protocol ValueFailure: Error, Equatable {}
-
-enum ColorValueFailure: ValueFailure {
-  /// The value is not a valid hexadecimal string.
-  case invalidHexString
-}
-
-enum DimensionValueFailure: ValueFailure {
-  /// The value does not represent a valid dimension.
-  case invalidStringValue
-
-  /// The value of the dimension is not a valid number.
-  case invalidValue
-}
-
-enum AliasValueFailure: ValueFailure {
-  /// The value of the alias is not a valid token reference.
-  case invalidReferenceSyntax
-}
