+++ Sources/DesignTokensCore/Errors/ValueFailure.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+/// A protocol defining requirements for a failure related to a token value.
+protocol ValueFailure: Error, Equatable {}
